requirements.txt
fastapi
uvicorn
pandas
requests
google-generativeai
